On Windows path.join uses '\' but a '/' is needed.
Update the example code because the path.join method uses an item separator that is OS dependent. On a Windows system path.join will result in a path like this:
```
foo\bar
```
In this case the prefix is not stripped, it is isn't found because the path returned by staticFileGlobs has UNIX separators. This example code should work on all OSs.